Fix wrong handling when faulting address is on PMD border
When PF happened on PMD border, an order 2 page was chosen but
then installed in the direction of the next highest page table.
This is wrong, since in the highest part of the next highest
page table, a order 8 should be installed, and only that.
Therefore before entering order choosing code, if the faulting
address is PMD border aligned, add 1 to it and then align it
upwards to get actual PMD block where we can operate. Like this,
the faulting address is no longer handled as if it were at the
beginning of the next page table, but as if it were at the end
of the current one.