Add functionality to allocate PMD-sized mTHP on stack
Add functionality, that if there is a page fault in a PMD block, which is not at the beginning of the stack, a PMD-sized mTHP is allocated. The reason for this is that if a PF happens, e.g., in the 2nd or 3rd-page table from the upper limit of the stack, it means that the first-page table is already full, or at least reserved for, e.g., a big array.